Create redirect from /league to specified url
Quick hard-link fix for now. There should be a way to get to get to the link without hard-coding it, or at least without hard-coding the port and the http://foosball.hq.axiosengineering.com, but this will work for now. The /league and /mobileapp (which is really the main app) are weirdly separate, so there's not a lot of crossover.